% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spatial_analysis.R
\name{generate_esda_plot}
\alias{generate_esda_plot}
\title{Generate ESDA Plots}
\usage{
generate_esda_plot(df, method)
}
\arguments{
\item{df}{A spatial dataframe (sf object) containing the outcome and geometry.}

\item{method}{A string indicating which of the 16 methods to execute.}
}
\description{
This function handles the 16 Exploratory Spatial Data Analysis methods.
}
