% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Read Golem Config}
\usage{
get_golem_config(str, config = Sys.getenv("GOLEM_CONFIG_ACTIVE", "default"))
}
\arguments{
\item{str}{The setting to read}

\item{config}{The configuration profile}
}
\description{
Read Golem Config
}
